+++ src/App.vue
@@ -23,7 +23,7 @@ import DocumentReview from "@/documentReview/Index.vue";
 import Portfolios from "@/portfolios/Index.vue";
 import Packages from "@/packages/Index.vue";
 import Home from "@/home/Index.vue";
-import ProvisionWorkflow from "@/home/ProvisionWorkflow.vue";
+import ProvisionWorkflow from "@/portfolios/provisioning/ProvisionWorkflow.vue";
 import CurrentUserStore from "./store/user";
 
 @Component({
